chore(cli): move commands to own directory
